# internal/api/openapi.json
{
  "components": {
    "schemas": {
...
      "IdentityResponse": {
        "description": "`GET /api/v1/identity` response: the caller's resolved identity.",
        "properties": {
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "groups": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "roles": {
            "description": "Snake_case role names (\"viewer\", \"developer\", \"operator\", \"admin\").",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "subject": {
            "type": "string"
          }
        },
        "required": [
          "subject",
          "groups",
          "roles"
        ],
        "type": "object"
      },
...
      "LoginIdentity": {
        "description": "The identity half of a login response.",
        "properties": {
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "subject": {
            "type": "string"
          }
        },
        "required": [
          "subject",
          "roles"
        ],
        "type": "object"
      },
      "LoginRequest": {
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "type": "object"
      },
      "LoginResponse": {
        "description": "`POST /api/v1/auth/login` success body. `token` is an opaque\n`mob_…` bearer credential (ADR-0011: stored, never signed).",
        "properties": {
          "expires_at": {
            "description": "Unix seconds after which the token no longer authenticates.",
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "identity": {
            "$ref": "#/components/schemas/LoginIdentity"
          },
          "token": {
            "type": "string"
          },
          "token_type": {
            "description": "Always \"bearer\".",
            "example": "bearer",
            "type": "string"
          }
        },
        "required": [
          "token",
          "token_type",
          "expires_at",
          "identity"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "FOSS control plane for Ray clusters. All endpoints take a Bearer JWT (OIDC). Note: the Ray Jobs API is NOT documented here — it is served by hostname, each registered cluster's hostname proxying Ray's own /api/jobs/ surface through the federating gateway. This spec covers Bifrost's own control-plane routes.",
    "license": {
      "name": "Apache-2.0"
    },
    "title": "Bifrost",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Logged in; opaque bearer token returned"
          },
          "401": {
            "description": "Invalid credentials (identical body for unknown user, wrong password, locked, or disabled)"
          }
        },
        "summary": "Username/password login (ADR-0011). Public (allowlisted); rate-limited\nby bcrypt cost and the 5-strikes/5-minute account lockout.",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/identity": {
      "get": {
        "operationId": "identity",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponse"
                }
              }
            },
            "description": "The caller's identity (the dev identity when auth is disabled)"
          },
          "401": {
            "description": "No/invalid token (only when auth is configured)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "\"Who am I\": the resolved identity for any authenticated caller. In dev\nmode (no validator AND no local auth) the auth middleware attaches no\nidentity, and this returns the specced dev identity so the\nunauthenticated dev loop renders the full console.",
        "tags": [
          "access"
        ]
      }
    },
...
  },
...
}